requirements.txt
pyusb>=1.1.0
pyserial>=3.4
docopt>=0.6.2
pylzma>=0.5.0
pycryptodome
qrcode
colorama~=0.4.4
enum34>=1.1.10
capstone>=4.0.2
keystone-engine
unicorn>=1.0.2
lxml>=4.6.1
future>=0.18.2
matplotlib>=3.3.3
numpy>=1.19.4
idna>=2.10
dnspython>=2.0.0
pyasn1>=0.4.8
ldap3>=2.8.1
asn1crypto>=1.4.0
pytz>=2020.4
qiling>=1.2
six>=1.15.0
requests>=2.25.0
grequests>=0.6.0

PyQt5~=5.14.1
coverage~=5.2
simplejson~=3.16.0